fix(license-detection): normalize match score to 0-100 range and fix rule_relevance type
Two bugs fixed:

1. LicenseMatch::score range inconsistency: aho_match and unknown_match
   produced scores in 0.0-1.0 while hash_match, spdx_lid, seq_match, and
   compute_match_score produced 0.0-100.0. The doc comment incorrectly
   stated 0.0-1.0. Fixed by scaling aho_match and unknown_match scores
   to 0.0-100.0 at construction, updating the doc comment, and updating
   all test assertions that used the old 0.0-1.0 convention.

2. rule_relevance usize→u8 truncation: models::Match::rule_relevance was
   Option<usize> but the internal LicenseMatch uses u8. The reverse
   conversion at reference_following.rs:1175 narrowed usize→u8 with no
   validation. Fixed by changing the model and output_schema types to
   Option<u8>, which is semantically correct for a 0-100 score and
   eliminates the narrowing cast.

+++ src/post_processing/mod.rs
@@ -608,7 +608,7 @@ mod tests {
     use crate::models::LineNumber;
     use crate::models::file_info::Match;
 
-    fn make_match(score: f64, coverage: Option<f64>, relevance: Option<usize>) -> Match {
+    fn make_match(score: f64, coverage: Option<f64>, relevance: Option<u8>) -> Match {
         Match {
             license_expression: "mit".to_string(),
             license_expression_spdx: "MIT".to_string(),
